Move cluster utility model implementation to its own file
